Fix policy for masterdata_admin and viewer roles

+++ db/schema.rb
@@ -57,8 +57,8 @@ ActiveRecord::Schema.define(version: 20180530081339) do
   end
 
   create_table "inquiry_inquiries_processors", id: false, force: :cascade do |t|
-    t.integer "inquiry_id", null: false
-    t.integer "processor_id", null: false
+    t.bigint "inquiry_id", null: false
+    t.bigint "processor_id", null: false
     t.index ["inquiry_id", "processor_id"], name: "index_inquiry_processor"
     t.index ["processor_id", "inquiry_id"], name: "index_processor_inquiry"
   end
